Adds global BlobDB concept. Refactorings, fixes.
Before we had blobs tied to individual tribledb.
However as we approach networking, it becomes clear that trible and
blobs need to be decomplected further and live in separate storages.
This way we can use tribles for message like realtime communication,
and collect blobs on a need to get basis, that uses smarter caching,
pulling and distributed protocols.